# reduce object allocations slightly