--- a/lib/faster_s3_url/shrine/storage.rb
+++ b/lib/faster_s3_url/shrine/storage.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 gem "shrine", "~> 3.0"
 require 'shrine/storage/s3'