Quite likely not enough to matter, but following trails from https://github.com/sciencehistory/scihist_digicoll/issues/2449